Add tunable params for pawn attacks on or near King.
Bump range of KING_ATTACK_SCALE to 90.

Restrict range of KING_ATTACK_COUNT_BOOST. Fix possible index overflow
in tuner when tuning these params.

Re-tune params with larger training set (and do not use regularization).

+++ src/search.cpp
@@ -85,10 +85,10 @@ static const int LMP_DEPTH=10;
 static const int LMP_MOVE_COUNT[11] = {3, 3, 5, 9, 15, 23, 33, 45, 59, 75, 93
 };
 
-static const int RAZOR_MARGIN[4] = { 0.9*PAWN_VALUE,
-                                     1.25*PAWN_VALUE,
-                                     3.0*PAWN_VALUE,
-                                     3.5*PAWN_VALUE };
+static const int RAZOR_MARGIN[4] = { int(0.9*PAWN_VALUE),
+                                     int(1.25*PAWN_VALUE),
+                                     int(3.0*PAWN_VALUE),
+                                     int(3.5*PAWN_VALUE) };
 
 static const int FUTILITY_MARGIN_BASE = (int)(1.01*PAWN_VALUE);
 static const int FUTILITY_MARGIN_SLOPE = (int)(0.772*PAWN_VALUE);
